+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Diskover Change Log
 
+## [1.5.0-rc29] = 2019-01-25
+### added
+- faster finddupes
+- worker bot warnings output for finddupes for any io/os exceptions
+- restoretimes config setting to dupescheck section in diskover.cfg.sample, copy to your config - setting to True will try to restore atime and mtime for any files which get opened from byte check and md5 (useful for cifs which does not work with noatime mount option)
+### changed
+- finddupes uses cpu count x 2 instead of 4 for threads
+
 ## [1.5.0-rc28] = 2019-01-19
 ### added
 - multiple es hosts can now be set in diskover.cfg elasticsearch section, see diskover.cfg.sample